requirements.txt
ccxt
ta
pandas
requests
python-dotenv